Function tests moved to unit dir, test function _common updated to accept coord name
